requirements.txt
Django
Pillow
django-simple-captcha
python-telegram-bot
PTable